Move the `Resource` type and supporting traits into generated code.
